persister_service/requirements.txt
pymongo==4.7.1
python-dotenv==1.0.0
kafka-python==2.1.1
